@@ -1,7 +1,7 @@
-cmake_minimum_required(VERSION 3.8 FATAL_ERROR)
-project(tga VERSION 1.0 LANGUAGES C CXX DESCRIPTION "Trainings Graphics API")
+cmake_minimum_required(VERSION 3.9.4 FATAL_ERROR)
+project(tga VERSION 2.0 LANGUAGES C CXX DESCRIPTION "Trainings Graphics API")
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 if(NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE Release)
@@ -10,7 +10,7 @@ endif()
 if(CMAKE_COMPILER_IS_GNUCXX)
     message(STATUS "GCC detected, adding compile flags")
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra")
-    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g -Og")
+    #set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g -Og")
     set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} -O3 -march=native")
 endif(CMAKE_COMPILER_IS_GNUCXX)
 if(WIN32)
@@ -69,7 +69,7 @@ add_subdirectory(src)
 
 option(TGA_BUILD_EXAMPLES on)
 
-#if(TGA_BUILD_EXAMPLES)
+if(TGA_BUILD_EXAMPLES)
 	message("Set TGA_BUILD_EXAMPLES to ${TGA_BUILD_EXAMPLES}: including subdirectory")
 	add_subdirectory(examples)
-#endif()
+endif()
