@@ -5,14 +5,14 @@
 
 int main()
 {
-    std::shared_ptr<tga::Interface> tgai = std::make_shared<tga::TGAVulkan>();
+    tga::Interface tgai{};
 
     // Use utility function to load shaders from file
-    tga::Shader vertexShader = tga::loadShader("shaders/drawIndirect_vert.spv", tga::ShaderType::vertex, tgai);
-    tga::Shader fragmentShader = tga::loadShader("shaders/drawIndirect_frag.spv", tga::ShaderType::fragment, tgai);
+    tga::Shader vertexShader = tga::loadShader("../shaders/drawIndirect_vert.spv", tga::ShaderType::vertex, tgai);
+    tga::Shader fragmentShader = tga::loadShader("../shaders/drawIndirect_frag.spv", tga::ShaderType::fragment, tgai);
 
-    auto [screenResX, screenResY] = tgai->screenResolution();
-    tga::Window window = tgai->createWindow({screenResX, screenResY});
+    auto [screenResX, screenResY] = tgai.screenResolution();
+    tga::Window window = tgai.createWindow({screenResX, screenResY});
 
     std::vector<tga::DrawIndirectCommand> indirectCommands{/*Draw red triangle*/
                                                            {/*vertexCount=*/3,
@@ -24,9 +24,12 @@ int main()
                                                            /*Draw blue rectangle*/
                                                            {6, 1, 6, 0}};
     // Create buffer storing the indirect draw commands
-    tga::Buffer indirectDrawBuffer =
-        tgai->createBuffer({tga::BufferUsage::indirect, tga::memoryAccess(indirectCommands),
-                            indirectCommands.size() * sizeof(tga::DrawIndirectCommand)});
+
+    auto stagingData = tgai.createStagingBuffer(
+        {indirectCommands.size() * sizeof(tga::DrawIndirectCommand), tga::memoryAccess(indirectCommands)});
+
+    tga::Buffer indirectDrawBuffer = tgai.createBuffer(
+        {tga::BufferUsage::indirect, indirectCommands.size() * sizeof(tga::DrawIndirectCommand), stagingData});
 
     struct Vertex {
         glm::vec2 position;
@@ -54,31 +57,28 @@ int main()
 
     tga::VertexLayout vertexLayout(sizeof(Vertex), {{offsetof(Vertex, position), tga::Format::r32g32_sfloat},
                                                     {offsetof(Vertex, color), tga::Format::r32g32b32_sfloat}});
-    tga::Buffer vertexBuffer = tgai->createBuffer(
-        {tga::BufferUsage::vertex, tga::memoryAccess(vertexBufferCPU), vertexBufferCPU.size() * sizeof(Vertex)});
 
-    tga::RenderPassInfo rpInfo{{vertexShader, fragmentShader}, window};
-    rpInfo.vertexLayout = vertexLayout;
+    auto vertexStaging =
+        tgai.createStagingBuffer({vertexBufferCPU.size() * sizeof(Vertex), tga::memoryAccess(vertexBufferCPU)});
+    tga::Buffer vertexBuffer =
+        tgai.createBuffer({tga::BufferUsage::vertex, vertexBufferCPU.size() * sizeof(Vertex), vertexStaging});
 
-    tga::RenderPass renderPass = tgai->createRenderPass(rpInfo);
+    auto rpInfo = tga::RenderPassInfo{vertexShader, fragmentShader, window}.setVertexLayout(vertexLayout);
+
+    tga::RenderPass renderPass = tgai.createRenderPass(rpInfo);
 
     tga::CommandBuffer cmdBuffer;
 
-    while (!tgai->windowShouldClose(window)) {
-        tgai->beginCommandBuffer(cmdBuffer);
+    while (!tgai.windowShouldClose(window)) {
+        auto nextFrame = tgai.nextFrame(window);
+        cmdBuffer = tga::CommandRecorder{tgai, cmdBuffer}
+                        .setRenderPass(renderPass, nextFrame)
+                        .bindVertexBuffer(vertexBuffer)
+                        .drawIndirect(indirectDrawBuffer, indirectCommands.size())
+                        .endRecording();
 
-        tgai->setRenderPass(renderPass, tgai->nextFrame(window));
-
-        tgai->bindVertexBuffer(vertexBuffer);
-
-        // Use indirect draw buffer for 3 draw commands
-        tgai->drawIndirect(indirectDrawBuffer,
-                           indirectCommands.size() /*offset=0, stride=sizeof(tga::DrawIndirectCommand)*/);
-
-        cmdBuffer = tgai->endCommandBuffer();
-
-        tgai->execute(cmdBuffer);
-        tgai->present(window);
+        tgai.execute(cmdBuffer);
+        tgai.present(window, nextFrame);
     }
     return 0;
 }
