@@ -7,26 +7,26 @@
 class Framework
 {
     public:
-    Framework():deltaTime(0.016),tgai(std::make_shared<tga::TGAVulkan>()){}
-    Framework(std::shared_ptr<tga::Interface> _tgai):deltaTime(0.016),tgai(_tgai){}
+    Framework():deltaTime(0.016),tgai(){}
     
     void run(uint32_t width = 0, uint32_t height = 0, bool vSyncOn = true)
     {
         if(!width || !height){
-            auto [w,h] = tgai->screenResolution();
+            auto [w,h] = tgai.screenResolution();
             width = w;
             height = h;
         }
         _frameworkWindowWidth = width;
         _frameworkWindowHeight = height;
         using namespace std::chrono_literals;
-        _frameworkWindow = tgai->createWindow({width,height,vSyncOn?tga::PresentMode::vsync:tga::PresentMode::immediate});
+        _frameworkWindow = tgai.createWindow({width,height,vSyncOn?tga::PresentMode::vsync:tga::PresentMode::immediate});
         OnCreate();
         auto beginTime = std::chrono::steady_clock::now();
-        while(!tgai->windowShouldClose(_frameworkWindow))
+        while(!tgai.windowShouldClose(_frameworkWindow))
         {
-            OnUpdate(tgai->nextFrame(_frameworkWindow));
-            tgai->present(_frameworkWindow);
+            auto nextFrame = tgai.nextFrame(_frameworkWindow);
+            OnUpdate(nextFrame);
+            tgai.present(_frameworkWindow,nextFrame);
             auto endTime = std::chrono::steady_clock::now();
             deltaTime = std::chrono::duration<double>(endTime-beginTime).count();
             beginTime = endTime;
@@ -41,7 +41,7 @@ class Framework
 
     protected:
     double deltaTime;
-    std::shared_ptr<tga::Interface> tgai;
+    tga::Interface tgai;
     tga::Window _frameworkWindow;
     uint32_t _frameworkWindowWidth,_frameworkWindowHeight;
 };
