@@ -2,169 +2,83 @@
 #include "tga/tga.hpp"
 #include "tga/tga_hash.hpp"
 #include "tga_vulkan_WSI.hpp"
-#include "tga_vulkan_util.hpp"
+#include "tga_vulkan_metadata.hpp"
 
 namespace tga
 {
-    /** \brief The Interface Implementation over the Vulkan API
-     */
-    class TGAVulkan final : public Interface {
-    public:
-        TGAVulkan();
-        ~TGAVulkan();
+/** \brief The Interface Implementation over the Vulkan API
+ */
+struct Interface::InternalState {
+    InternalState();
+    // Vulkan Stuff
+    VulkanWSI wsi;
+    vk::Instance instance;
+    vk::DebugUtilsMessengerEXT debugger;
+    vk::PhysicalDevice pDevice;
+    uint32_t hostMemoryIndex;
+    uint32_t deviceMemoryIndex;
+    uint32_t renderQueueFamily;
+    vk::Device device;
+    vk::Queue renderQueue;
+    vk::CommandPool cmdPool;
 
-        Shader createShader(const ShaderInfo &shaderInfo) override;
-        Buffer createBuffer(const BufferInfo &bufferInfo) override;
-        Texture createTexture(const TextureInfo &textureInfo) override;
-        Window createWindow(const WindowInfo &windowInfo) override;
-        InputSet createInputSet(const InputSetInfo &inputSetInfo) override;
-        RenderPass createRenderPass(const RenderPassInfo &renderPassInfo) override;
+    vkData::Buffer allocateBuffer(vk::DeviceSize size, vk::BufferUsageFlags usage, vk::MemoryPropertyFlags properties);
+    std::tuple<vk::ImageType, vk::ImageViewType, vk::ImageCreateFlags> determineImageTypeInfo(
+        const TextureInfo& textureInfo);
+    std::tuple<vk::Extent3D, uint32_t> determineImageDimensions(const TextureInfo& textureInfo);
+    std::pair<vk::ImageTiling, vk::ImageUsageFlags> determineImageFeatures(vk::Format& format);
+    vk::Format findDepthFormat();
+    vkData::DepthBuffer createDepthBuffer(uint32_t width, uint32_t height);
+    vk::RenderPass makeRenderPass(std::vector<vk::Format> const& colorFormats, ClearOperation clearOps,
+                                  vk::ImageLayout layout);
+    std::vector<vk::DescriptorSetLayout> decodeInputLayout(InputLayout const& inputLayout);
+    vk::Pipeline makeGraphicsPipeline(const RenderPassInfo& renderPassInfo, vk::PipelineLayout pipelineLayout,
+                                      vk::RenderPass renderPass);
+    std::pair<vk::Pipeline, vk::PipelineBindPoint> makePipeline(const RenderPassInfo& renderPassInfo,
+                                                                vk::PipelineLayout pipelineLayout,
+                                                                vk::RenderPass renderPass);
 
-        void beginCommandBuffer() override;
-        void beginCommandBuffer(CommandBuffer cmdBuffer) override;
-        void setRenderPass(RenderPass renderPass, uint32_t frambufferIndex) override;
-        void bindVertexBuffer(Buffer buffer) override;
-        void bindIndexBuffer(Buffer buffer) override;
-        void bindInputSet(InputSet inputSet) override;
-        void draw(uint32_t vertexCount, uint32_t firstVertex, uint32_t instanceCount = 1,
-                  uint32_t firstInstance = 0) override;
-        void drawIndexed(uint32_t indexCount, uint32_t firstIndex, uint32_t vertexOffset, uint32_t instanceCount = 1,
-                         uint32_t firstInstance = 0) override;
-        void drawIndirect(Buffer buffer, uint32_t drawCount, size_t offset = 0,
-                          uint32_t stride = sizeof(tga::DrawIndirectCommand)) override;
-        void drawIndexedIndirect(Buffer buffer, uint32_t drawCount, size_t offset = 0,
-                                 uint32_t stride = sizeof(tga::DrawIndexedIndirectCommand)) override;
-        void dispatch(uint32_t groupCountX, uint32_t groupCountY, uint32_t groupCountZ) override;
-        CommandBuffer endCommandBuffer() override;
-        void execute(CommandBuffer commandBuffer) override;
+    // vk::CommandBuffer beginOneTimeCmdBuffer(vk::CommandPool &cmdPool);
+    // void endOneTimeCmdBuffer(vk::CommandBuffer &cmdBuffer, vk::CommandPool &cmdPool, vk::Queue &submitQueue);
 
-        void updateBuffer(Buffer buffer, uint8_t const *data, size_t dataSize, uint32_t offset) override;
-        std::vector<uint8_t> readback(Buffer buffer) override;
-        std::vector<uint8_t> readback(Texture texture) override;
+    // void fillBuffer(size_t size, const uint8_t *data, uint32_t offset, vk::Buffer target);
+    // void fillTexture(size_t size, const uint8_t *data, vk::Extent3D extent, uint32_t layers, vk::Image target);
 
-        /** \copydoc Interface::backbufferCount(Window window)
-         */
-        uint32_t backbufferCount(Window window) override;
+    // Bookkeeping
+    // Note: Windows are stored in WSI
+    std::vector<vkData::Shader> shaders;
+    std::vector<vkData::Buffer> buffers;
+    std::vector<vkData::StagingBuffer> stagingBuffers;
+    std::vector<vkData::Texture> textures;
+    std::vector<vkData::InputSet> inputSets;
+    std::vector<vkData::RenderPass> renderPasses;
+    std::vector<vkData::ComputePass> computePasses;
+    std::vector<vkData::CommandBuffer> commandBuffers;
+    std::vector<vkData::ext::AccelerationStructure> acclerationStructures;
 
-        /** \copydoc Interface::nextFrame(Window window)
-         */
-        uint32_t nextFrame(Window window) override;
+    vkData::Shader& getData(Shader);
+    vkData::Buffer& getData(Buffer);
+    vkData::StagingBuffer& getData(StagingBuffer);
+    vkData::Texture& getData(Texture);
+    vkData::Window& getData(Window);
+    vkData::InputSet& getData(InputSet);
+    vkData::RenderPass& getData(RenderPass);
+    vkData::ComputePass& getData(ComputePass);
+    vkData::CommandBuffer& getData(CommandBuffer);
+    vkData::ext::AccelerationStructure& getData(ext::TopLevelAccelerationStructure);
+    vkData::ext::AccelerationStructure& getData(ext::BottomLevelAccelerationStructure);
 
-        /** \copydoc Interface::pollEvents(Window window)
-         */
-        void pollEvents(Window window) override;
+    struct RecordingData {
+        vk::CommandBuffer cmdBuffer;
+        RenderPass renderPass;
+    } currentRecording;
 
-        /** \copydoc Interface::present(Window window)
-         */
-        void present(Window window) override;
-
-        /** \copydoc Interface::setWindowTitle(Window window, const std::string &title)
-         */
-        void setWindowTitle(Window window, const std::string &title) override;
-
-        /** \copydoc Interface::windowShouldClose(Window window)
-         */
-        bool windowShouldClose(Window window) override;
-
-        /** \copydoc Interface::keyDown(Window window)
-         */
-        bool keyDown(Window window, Key key) override;
-
-        /** \copydoc Interface::mousePosition(Window window)
-         */
-        std::pair<int, int> mousePosition(Window window) override;
-
-        /** \copydoc Interface::screenResolution()
-         */
-        std::pair<uint32_t, uint32_t> screenResolution() override;
-
-        void free(Shader shader) override;
-        void free(Buffer buffer) override;
-        void free(Texture texture) override;
-        void free(Window window) override;
-        void free(InputSet inputSet) override;
-        void free(RenderPass renderPass) override;
-        void free(CommandBuffer commandBuffer) override;
-
-    private:
-        // Vulkan Stuff
-        VulkanWSI wsi;
-        vk::Instance instance;
-        vk::DebugUtilsMessengerEXT debugger;
-        vk::PhysicalDevice pDevice;
-        QueueIndices queueIndices;
-        vk::Device device;
-        vk::Queue graphicsQueue;
-        vk::Queue transferQueue;
-        vk::CommandPool transferCmdPool;
-        vk::CommandPool graphicsCmdPool;
-
-        const std::vector<const char *> getInstanceExtentensions();
-        const std::vector<const char *> getDeviceExtentensions();
-        const std::vector<const char *> getLayers();
-        vk::PhysicalDeviceFeatures getDeviceFeatures();
-        uint32_t findQueueFamily(vk::QueueFlags mask, vk::QueueFlags flags);
-        QueueIndices findQueueFamilies();
-
-        vk::Instance createInstance();
-        vk::DebugUtilsMessengerEXT createDebugger();
-        vk::PhysicalDevice choseGPU();
-        vk::Device createDevice();
-        vk::CommandPool createCommandPool(uint32_t queueFamily,
-                                          vk::CommandPoolCreateFlags flags = vk::CommandPoolCreateFlags());
-        uint32_t findMemoryType(uint32_t typeFilter, vk::MemoryPropertyFlags properties);
-        Buffer_TV allocateBuffer(vk::DeviceSize size, vk::BufferUsageFlags usage, vk::MemoryPropertyFlags properties);
-        std::tuple<vk::ImageType, vk::ImageViewType, vk::ImageCreateFlags> determineImageTypeInfo(
-            const TextureInfo &textureInfo);
-        std::tuple<vk::Extent3D, uint32_t> determineImageDimensions(const TextureInfo &textureInfo);
-        std::pair<vk::ImageTiling, vk::ImageUsageFlags> determineImageFeatures(vk::Format &format);
-        vk::Format findDepthFormat();
-        DepthBuffer_TV createDepthBuffer(uint32_t width, uint32_t height);
-        vk::RenderPass makeRenderPass(std::vector<vk::Format> const& colorFormats, ClearOperation clearOps, vk::ImageLayout layout);
-        std::vector<vk::DescriptorSetLayout> decodeInputLayout(const InputLayout &inputLayout);
-        vk::Pipeline makeGraphicsPipeline(const RenderPassInfo &renderPassInfo, vk::PipelineLayout pipelineLayout,
-                                          vk::RenderPass renderPass);
-        std::pair<vk::Pipeline, vk::PipelineBindPoint> makePipeline(const RenderPassInfo &renderPassInfo,
-                                                                    vk::PipelineLayout pipelineLayout,
-                                                                    vk::RenderPass renderPass);
-
-        vk::CommandBuffer beginOneTimeCmdBuffer(vk::CommandPool &cmdPool);
-        void endOneTimeCmdBuffer(vk::CommandBuffer &cmdBuffer, vk::CommandPool &cmdPool, vk::Queue &submitQueue);
-
-        void fillBuffer(size_t size, const uint8_t *data, uint32_t offset, vk::Buffer target);
-        void transitionImageLayout(vk::CommandBuffer cmdBuffer, vk::Image image, vk::ImageLayout oldLayout,
-                                   vk::ImageLayout newLayout);
-        void fillTexture(size_t size, const uint8_t *data, vk::Extent3D extent, uint32_t layers, vk::Image target);
-
-        // Convertes
-        vk::BufferUsageFlags determineBufferFlags(tga::BufferUsage usage);
-        vk::Format determineImageFormat(tga::Format format);
-        std::tuple<vk::Filter, vk::SamplerAddressMode> determineSamplerInfo(const TextureInfo &textureInfo);
-        vk::ShaderStageFlagBits determineShaderStage(tga::ShaderType shaderType);
-        std::vector<vk::VertexInputAttributeDescription> determineVertexAttributes(
-            const std::vector<VertexAttribute> &attributes);
-        vk::PipelineRasterizationStateCreateInfo determineRasterizerState(const RasterizerConfig &config);
-        vk::CompareOp determineDepthCompareOp(CompareOperation compareOperation);
-        vk::BlendFactor determineBlendFactor(BlendFactor blendFactor);
-        vk::PipelineColorBlendAttachmentState determineColorBlending(const PerPixelOperations &config);
-        vk::DescriptorType determineDescriptorType(tga::BindingType bindingType);
-        vk::AccessFlags layoutToAccessFlags(vk::ImageLayout layout);
-        vk::PipelineStageFlags layoutToPipelineStageFlags(vk::ImageLayout layout);
-        vk::PipelineStageFlags accessToPipelineStageFlags(vk::AccessFlags accessFlags);
-
-        // Bookkeeping
-        std::unordered_map<Shader, Shader_TV> shaders;
-        std::unordered_map<Buffer, Buffer_TV> buffers;
-        std::unordered_map<Texture, Texture_TV> textures;
-        std::unordered_map<InputSet, InputSet_TV> inputSets;
-        std::unordered_map<RenderPass, RenderPass_TV> renderPasses;
-        std::unordered_map<CommandBuffer, CommandBuffer_TV> commandBuffers;
-        std::unordered_map<Texture, DepthBuffer_TV> textureDepthBuffers;
-        std::unordered_map<Window, DepthBuffer_TV> windowDepthBuffers;
-
-        struct RecordingData {
-            vk::CommandBuffer cmdBuffer;
-            RenderPass renderPass;
-        } currentRecording;
-    };
+    void free(Shader shader);
+    void free(Buffer buffer);
+    void free(Texture texture);
+    void free(Window window);
+    void free(InputSet inputSet);
+    void free(RenderPass renderPass);
+    void free(CommandBuffer commandBuffer);
+};
 }  // namespace tga
